basic fixup of warnings

+++ src/v_video.c
@@ -2037,8 +2037,8 @@ Unoptimized version
 
 		float cosma = cos(f_angle);
 
-		float xst = (INT32)round((cosma * -hwidth - sinma * -hheight) + hwidth);
-		float yst = (INT32)round((sinma * -hwidth + cosma * -hheight) + hheight);
+		float xst = round((cosma * -hwidth - sinma * -hheight) + hwidth);
+		float yst = round((sinma * -hwidth + cosma * -hheight) + hheight);
 
 		// Fills the empty space with a solid color from palette index
 		memset(screens[4], (UINT8)(31), vid.width*vid.height*vid.bpp);
